Move nullifier batch insertion logic
