+++ yarn-project/circuits.js/src/structs/rollup/base_rollup.ts
@@ -162,7 +162,7 @@ export class BaseRollupInputs {
     public startPublicDataTreeRoot: Fr,
 
     /**
-     * The nullifiers which need to be udpated to perform the batch insertion of the new nullifiers.
+     * The nullifiers which need to be updated to perform the batch insertion of the new nullifiers.
      * See `performBaseRollupBatchInsertionProofs` function for more details.
      */
     public lowNullifierLeafPreimages: NullifierLeafPreimage[],
