+++ yarn-project/sequencer-client/src/block_builder/solo_block_builder.ts
@@ -25,9 +25,8 @@ import {
   VerificationKey,
 } from '@aztec/circuits.js';
 import { computeContractLeaf } from '@aztec/circuits.js/abis';
-import { LeafData, SiblingPath } from '@aztec/merkle-tree';
 import { MerkleTreeId, ContractData, L2Block, PublicDataWrite } from '@aztec/types';
-import { MerkleTreeOperations } from '@aztec/world-state';
+import { LowNullifierWitnessData, MerkleTreeOperations } from '@aztec/world-state';
 import chunk from 'lodash.chunk';
 import flatMap from 'lodash.flatmap';
 import times from 'lodash.times';
@@ -39,7 +38,7 @@ import { ProcessedTx } from '../sequencer/processed_tx.js';
 import { BlockBuilder } from './index.js';
 import { Fr } from '@aztec/foundation/fields';
 import { createDebugLogger } from '@aztec/foundation/log';
-import { toBigIntBE, toBufferBE } from '@aztec/foundation/bigint-buffer';
+import { toBigIntBE } from '@aztec/foundation/bigint-buffer';
 import { toFriendlyJSON } from '@aztec/circuits.js/utils';
 import { AllowedTreeNames, OutputWithTreeSnapshot } from './types.js';
 
@@ -53,31 +52,6 @@ const FUTURE_NUM = 0;
 
 // Denotes fields that should be deleted
 const DELETE_FR = new Fr(0n);
-
-/**
- * All of the data required for the circuit compute and verify nullifiers.
- */
-export interface LowNullifierWitnessData {
-  /**
-   * Preimage of the low nullifier that proves non membership.
-   */
-  preimage: NullifierLeafPreimage;
-  /**
-   * Sibling path to prove membership of low nullifier.
-   */
-  siblingPath: SiblingPath;
-  /**
-   * The index of low nullifier.
-   */
-  index: bigint;
-}
-
-// Pre-compute empty nullifier witness
-const EMPTY_LOW_NULLIFIER_WITNESS: LowNullifierWitnessData = {
-  preimage: NullifierLeafPreimage.empty(),
-  index: 0n,
-  siblingPath: new SiblingPath(Array(NULLIFIER_TREE_HEIGHT).fill(toBufferBE(0n, 32))),
-};
 
 /**
  * Builds an L2 block out of a set of ProcessedTx's,
@@ -598,246 +572,6 @@ export class SoloBlockBuilder implements BlockBuilder {
     return fullSiblingPath.data.slice(subtreeHeight).map(b => Fr.fromBuffer(b));
   }
 
-  /* eslint-disable jsdoc/require-description-complete-sentence */
-  /* The following doc block messes up with complete-sentence, so we just disable it */
-
-  /**
-   *
-   * Each base rollup needs to provide non membership / inclusion proofs for each of the nullifier.
-   * This method will return membership proofs and perform partial node updates that will
-   * allow the circuit to incrementally update the tree and perform a batch insertion.
-   *
-   * This offers massive circuit performance savings over doing incremental insertions.
-   *
-   * A description of the algorithm can be found here: https://colab.research.google.com/drive/1A0gizduSi4FIiIJZ8OylwIpO9-OTqV-R
-   *
-   * WARNING: This function has side effects, it will insert values into the tree.
-   *
-   * Assumptions:
-   * 1. There are 8 nullifiers provided and they are either unique or empty. (denoted as 0)
-   * 2. If kc 0 has 1 nullifier, and kc 1 has 3 nullifiers the layout will assume to be the sparse
-   *   nullifier layout: [kc0-0, 0, 0, 0, kc1-0, kc1-1, kc1-2, 0]
-   *
-   * Algorithm overview
-   *
-   * In general, if we want to batch insert items, we first to update their low nullifier to point to them,
-   * then batch insert all of the values as at once in the final step.
-   * To update a low nullifier, we provide an insertion proof that the low nullifier currently exists to the
-   * circuit, then update the low nullifier.
-   * Updating this low nullifier will in turn change the root of the tree. Therefore future low nullifier insertion proofs
-   * must be given against this new root.
-   * As a result, each low nullifier membership proof will be provided against an intermediate tree state, each with differing
-   * roots.
-   *
-   * This become tricky when two items that are being batch inserted need to update the same low nullifier, or need to use
-   * a value that is part of the same batch insertion as their low nullifier. In this case a zero low nullifier path is given
-   * to the circuit, and it must determine from the set of batch inserted values if the insertion is valid.
-   *
-   * The following example will illustrate attempting to insert 2,3,20,19 into a tree already containing 0,5,10,15
-   *
-   * The example will explore two cases. In each case the values low nullifier will exist within the batch insertion,
-   * One where the low nullifier comes before the item in the set (2,3), and one where it comes after (20,19).
-   *
-   * The original tree:                       Pending insertion subtree
-   *
-   *  index     0       2       3       4         -       -       -       -
-   *  -------------------------------------      ----------------------------
-   *  val       0       5      10      15         -       -       -       -
-   *  nextIdx   1       2       3       0         -       -       -       -
-   *  nextVal   5      10      15       0         -       -       -       -
-   *
-   *
-   * Inserting 2: (happy path)
-   * 1. Find the low nullifier (0) - provide inclusion proof
-   * 2. Update its pointers
-   * 3. Insert 2 into the pending subtree
-   *
-   *  index     0       2       3       4         5       -       -       -
-   *  -------------------------------------      ----------------------------
-   *  val       0       5      10      15         2       -       -       -
-   *  nextIdx   5       2       3       0         2       -       -       -
-   *  nextVal   2      10      15       0         5       -       -       -
-   *
-   * Inserting 3: The low nullifier exists within the insertion current subtree
-   * 1. When looking for the low nullifier for 3, we will receive 0 again as we have not inserted 2 into the main tree
-   *    This is problematic, as we cannot use either 0 or 2 as our inclusion proof.
-   *    Why cant we?
-   *      - Index 0 has a val 0 and nextVal of 2. This is NOT enough to prove non inclusion of 2.
-   *      - Our existing tree is in a state where we cannot prove non inclusion of 3.
-   *    We do not provide a non inclusion proof to out circuit, but prompt it to look within the insertion subtree.
-   * 2. Update pending insertion subtree
-   * 3. Insert 3 into pending subtree
-   *
-   * (no inclusion proof provided)
-   *  index     0       2       3       4         5       6       -       -
-   *  -------------------------------------      ----------------------------
-   *  val       0       5      10      15         2       3       -       -
-   *  nextIdx   5       2       3       0         6       2       -       -
-   *  nextVal   2      10      15       0         3       5       -       -
-   *
-   * Inserting 20: (happy path)
-   * 1. Find the low nullifier (15) - provide inculsion proof
-   * 2. Update its pointers
-   * 3. Insert 20 into the pending subtree
-   *
-   *  index     0       2       3       4         5       6       7       -
-   *  -------------------------------------      ----------------------------
-   *  val       0       5      10      15         2       3      20       -
-   *  nextIdx   5       2       3       7         6       2       0       -
-   *  nextVal   2      10      15      20         3       5       0       -
-   *
-   * Inserting 19:
-   * 1. In this case we can find a low nullifier, but we are updating a low nullifier that has already been updated
-   *    We can provide an inclusion proof of this intermediate tree state.
-   * 2. Update its pointers
-   * 3. Insert 19 into the pending subtree
-   *
-   *  index     0       2       3       4         5       6       7       8
-   *  -------------------------------------      ----------------------------
-   *  val       0       5      10      15         2       3      20       19
-   *  nextIdx   5       2       3       8         6       2       0       7
-   *  nextVal   2      10      15      19         3       5       0       20
-   *
-   * Perform subtree insertion
-   *
-   *  index     0       2       3       4       5       6       7       8
-   *  ---------------------------------------------------------------------
-   *  val       0       5      10      15       2       3      20       19
-   *  nextIdx   5       2       3       8       6       2       0       7
-   *  nextVal   2      10      15      19       3       5       0       20
-   *
-   * TODO: this implementation will change once the zero value is changed from h(0,0,0). Changes incoming over the next sprint
-   * @param leaves - Values to insert into the tree
-   * @returns The witness data for the leaves to be updated when inserting the new ones.
-   */
-  public async performBaseRollupBatchInsertionProofs(
-    leaves: Buffer[],
-  ): Promise<[LowNullifierWitnessData[], Fr[]] | [undefined, Fr[]]> {
-    /* eslint-enable */
-
-    // Keep track of touched low nullifiers
-    const touched = new Map<number, bigint[]>();
-
-    // Accumulators
-    const lowNullifierWitnesses: LowNullifierWitnessData[] = [];
-    const pendingInsertionSubtree: NullifierLeafPreimage[] = [];
-
-    // Start info
-    const dbInfo = await this.db.getTreeInfo(MerkleTreeId.NULLIFIER_TREE);
-    const startInsertionIndex: bigint = dbInfo.size;
-
-    // Get insertion path for each leaf
-    for (let i = 0; i < leaves.length; i++) {
-      const newValue = toBigIntBE(leaves[i]);
-
-      // Keep space and just insert zero values
-      if (newValue === 0n) {
-        pendingInsertionSubtree.push(NullifierLeafPreimage.empty());
-        lowNullifierWitnesses.push(EMPTY_LOW_NULLIFIER_WITNESS);
-        continue;
-      }
-
-      const indexOfPrevious = await this.db.getPreviousValueIndex(MerkleTreeId.NULLIFIER_TREE, newValue);
-
-      // If a touched node has a value that is less greater than the current value
-      const prevNodes = touched.get(indexOfPrevious.index);
-      if (prevNodes && prevNodes.some(v => v < newValue)) {
-        // check the pending low nullifiers for a low nullifier that works
-        // This is the case where the next value is less than the pending
-        for (let j = 0; j < pendingInsertionSubtree.length; j++) {
-          if (pendingInsertionSubtree[j].leafValue.isZero()) continue;
-
-          if (
-            pendingInsertionSubtree[j].leafValue.value < newValue &&
-            (pendingInsertionSubtree[j].nextValue.value > newValue || pendingInsertionSubtree[j].nextValue.isZero())
-          ) {
-            // add the new value to the pending low nullifiers
-            const currentLeafLowNullifier = new NullifierLeafPreimage(
-              new Fr(newValue),
-              pendingInsertionSubtree[j].nextValue,
-              Number(pendingInsertionSubtree[j].nextIndex),
-            );
-
-            pendingInsertionSubtree.push(currentLeafLowNullifier);
-
-            // Update the pending low nullifier to point at the new value
-            pendingInsertionSubtree[j].nextValue = new Fr(newValue);
-            pendingInsertionSubtree[j].nextIndex = Number(startInsertionIndex) + i;
-
-            break;
-          }
-        }
-
-        // Any node updated in this space will need to calculate its low nullifier from a previously inserted value
-        lowNullifierWitnesses.push(EMPTY_LOW_NULLIFIER_WITNESS);
-      } else {
-        // Update the touched mapping
-        if (prevNodes) {
-          prevNodes.push(newValue);
-          touched.set(indexOfPrevious.index, prevNodes);
-        } else {
-          touched.set(indexOfPrevious.index, [newValue]);
-        }
-
-        // get the low nullifier
-        const lowNullifier = await this.db.getLeafData(MerkleTreeId.NULLIFIER_TREE, indexOfPrevious.index);
-        if (lowNullifier === undefined) {
-          return [
-            undefined,
-            await this.getSubtreeSiblingPath(MerkleTreeId.NULLIFIER_TREE, BaseRollupInputs.NULLIFIER_SUBTREE_HEIGHT),
-          ];
-        }
-
-        const lowNullifierPreimage = new NullifierLeafPreimage(
-          new Fr(lowNullifier.value),
-          new Fr(lowNullifier.nextValue),
-          Number(lowNullifier.nextIndex),
-        );
-        const siblingPath = await this.db.getSiblingPath(MerkleTreeId.NULLIFIER_TREE, BigInt(indexOfPrevious.index));
-
-        // Update the running paths
-        const witness: LowNullifierWitnessData = {
-          preimage: lowNullifierPreimage,
-          index: BigInt(indexOfPrevious.index),
-          siblingPath: siblingPath,
-        };
-        lowNullifierWitnesses.push(witness);
-
-        // The low nullifier the inserted value will have
-        const currentLeafLowNullifier = new NullifierLeafPreimage(
-          new Fr(newValue),
-          new Fr(lowNullifier.nextValue),
-          Number(lowNullifier.nextIndex),
-        );
-        pendingInsertionSubtree.push(currentLeafLowNullifier);
-
-        // Update the old low nullifier
-        lowNullifier.nextValue = newValue;
-        lowNullifier.nextIndex = startInsertionIndex + BigInt(i);
-
-        await this.db.updateLeaf(MerkleTreeId.NULLIFIER_TREE, lowNullifier, BigInt(indexOfPrevious.index));
-      }
-    }
-
-    const newNullifiersSubtreeSiblingPath = await this.getSubtreeSiblingPath(
-      MerkleTreeId.NULLIFIER_TREE,
-      BaseRollupInputs.NULLIFIER_SUBTREE_HEIGHT,
-    );
-
-    // Perform batch insertion of new pending values
-    for (let i = 0; i < pendingInsertionSubtree.length; i++) {
-      const asLeafData: LeafData = {
-        value: pendingInsertionSubtree[i].leafValue.value,
-        nextValue: pendingInsertionSubtree[i].nextValue.value,
-        nextIndex: BigInt(pendingInsertionSubtree[i].nextIndex),
-      };
-
-      await this.db.updateLeaf(MerkleTreeId.NULLIFIER_TREE, asLeafData, startInsertionIndex + BigInt(i));
-    }
-
-    return [lowNullifierWitnesses, newNullifiersSubtreeSiblingPath];
-  }
-
   protected async processPublicDataUpdateRequests(tx: ProcessedTx) {
     const newPublicDataUpdateRequestsSiblingPaths: MembershipWitness<typeof PUBLIC_DATA_TREE_HEIGHT>[] = [];
     for (const publicDataUpdateRequest of tx.data.end.publicDataUpdateRequests) {
@@ -914,7 +648,7 @@ export class SoloBlockBuilder implements BlockBuilder {
     // Update the nullifier tree, capturing the low nullifier info for each individual operation
     const newNullifiers = [...left.data.end.newNullifiers, ...right.data.end.newNullifiers];
 
-    const [nullifierWitnesses, newNullifiersSubtreeSiblingPath] = await this.performBaseRollupBatchInsertionProofs(
+    const [nullifierWitnesses, newNullifiersSubtreeSiblingPath] = await this.db.performBaseRollupBatchInsertionProofs(
       newNullifiers.map(fr => fr.toBuffer()),
     );
     if (nullifierWitnesses === undefined) {
